refactor: modularize game screen widget
